Raspberry pi/requirements.txt
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
Cython==0.29.28
Deprecated==1.2.13
docopt==0.6.2
firebase-admin==5.3.0
gcloud==0.18.3
gensim==4.2.0
google-api-core==2.8.2
google-api-python-client==2.58.0
google-auth==2.11.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.1
google-cloud-storage==2.5.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.47.0
grpcio-status==1.47.0
httplib2==0.20.4
idna==3.3
jwcrypto==1.3.1
msgpack==1.0.4
numpy==1.23.2
oauth2client==4.1.3
pipreqs==0.4.11
proto-plus==1.22.0
protobuf==4.21.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
pycryptodomex==3.15.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-dotenv==0.20.0
python-jwt==3.3.3
requests==2.28.1
requests-toolbelt==0.9.1
rsa==4.9
scipy==1.9.0
six==1.16.0
smart-open==6.1.0
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
yarg==0.1.9
